+++ nonsoolmateServer/src/main/java/com/nonsoolmate/nonsoolmateServer/domain/auth/controller/dto/response/MemberReissueResponseDTO.java
@@ -3,10 +3,9 @@ package com.nonsoolmate.nonsoolmateServer.domain.auth.controller.dto.response;
 import com.nonsoolmate.nonsoolmateServer.external.oauth.service.vo.enums.AuthType;
 import jakarta.validation.constraints.NotNull;
 
-public record MemberReissueResponseDTO(@NotNull Long memberId, @NotNull AuthType authType,
-                                       @NotNull String accessToken, @NotNull String refreshToken) {
-    public static MemberReissueResponseDTO of(Long memberId, AuthType authType, String accessToken,
+public record MemberReissueResponseDTO(@NotNull Long memberId, @NotNull String accessToken, @NotNull String refreshToken) {
+    public static MemberReissueResponseDTO of(Long memberId, String accessToken,
                                            String refreshToken) {
-        return new MemberReissueResponseDTO(memberId, authType, accessToken, refreshToken);
+        return new MemberReissueResponseDTO(memberId, accessToken, refreshToken);
     }
 }
