+++ nonsoolmateServer/src/main/java/com/nonsoolmate/nonsoolmateServer/domain/member/entity/enums/Role.java
@@ -1,5 +1,5 @@
 package com.nonsoolmate.nonsoolmateServer.domain.member.entity.enums;
 
 public enum Role {
-    ROLE_ADMIN, ROLE_USER
+    ADMIN, USER
 }
