--- nonsoolmateServer/src/main/java/com/nonsoolmate/nonsoolmateServer/global/auth/exception/AuthSucessType.java
@@ -1,4 +1,0 @@
-package com.nonsoolmate.nonsoolmateServer.global.auth.exception;
-
-public class AuthSucessType {
-}
