+++ nonsoolmateServer/src/main/java/com/nonsoolmate/nonsoolmateServer/global/security/service/MemberAuthService.java
@@ -1,12 +1,11 @@
-package com.nonsoolmate.nonsoolmateServer.global.jwt.service;
+package com.nonsoolmate.nonsoolmateServer.global.security.service;
 
 import com.nonsoolmate.nonsoolmateServer.domain.member.entity.Member;
 import com.nonsoolmate.nonsoolmateServer.domain.member.repository.MemberRepository;
 
+import com.nonsoolmate.nonsoolmateServer.global.security.CustomAuthUser;
 import com.nonsoolmate.nonsoolmateServer.global.jwt.utils.PasswordUtil;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.User.UserBuilder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
@@ -23,14 +22,6 @@ public class MemberAuthService implements UserDetailsService {
 
         String password = PasswordUtil.generateRandomPassword();
 
-        UserBuilder password1 = User.builder()
-                .username(member.getEmail())
-                .password(password);
-
-        return org.springframework.security.core.userdetails.User.builder()
-                .username(member.getEmail())
-                .password(password)
-                .roles(member.getRole().name())
-                .build();
+        return new CustomAuthUser(member, member.getRole());
     }
 }
